QNX: W/A for virtio-input over shmem
For virtio-input devices, writing to config area is defined as selecting
config pages. But in virtio over shmem implement, this operation is not
supported. The W/A is to update config area upon config selection in BE,
and delay a short while after cfg write in FE. This delay does not reduce
performance as it only happens during init stage.

Test done:
Using keyboard/mouse/touchscreen over virtio-input on QNX

Tracked-On: OAM-125247
Tracked-On: OAM-127456
Tracked-On: OAM-127455
Tracked-On: OAM-125249
Signed-off-by: Zhao, Shirley <shirley.zhao@intel.com>
Signed-off-by: Wu Zhou <wu.zhou@intel.com>