i915_ag: Control out of tree module loading using build flag
When kernel is built with bazel, i915_ag is built as module and it is
loaded durig init fs stage. Loading it as part of kernel results in
longer boot time due to loading it as part of kernel.

Fix the issue by enabling load of out of tree i915_ag module only when
ENABLE_I915_OOT_MODULE_LOADING is defined.

Change-Id: I8884176038c7b2c216684dd8a162c90df9778645
Signed-off-by: Jeevaka Prabu Badrappan <jeevaka.badrappan@intel.com>